rename containers to pages folder
