+++ client/src/pages/PaintingDetailsPage/index.js
