+++ client/src/pages/PaintingDetailsPage/style.css
